@@ -8,84 +8,110 @@
         "build": "cross-env NODE_ENV=production gulp",
         "build:dev": "cross-env NODE_ENV=development gulp",
         "start": "cross-env NODE_ENV=production gulp start",
-        "start:dev": "cross-env NODE_ENV=development gulp start"
+        "start:dev": "cross-env NODE_ENV=development gulp start",
+        "jest": "cross-env NODE_ENV=development gulp jest"
     },
     "license": "MIT",
     "dependencies": {
-        "@babel/runtime": "7.12.5",
-        "@fortawesome/react-fontawesome": "^0.1.14",
+        "@babel/runtime": "7.12.13",
+        "@fortawesome/fontawesome-svg-core": "1.2.34",
+        "@fortawesome/free-solid-svg-icons": "5.15.2",
+        "@fortawesome/react-fontawesome": "0.1.14",
+        "@types/immutable": "3.8.7",
         "connected-react-router": "6.8.0",
         "history": "4.10.1",
+        "immutable": "4.0.0-rc.12",
         "ramda": "0.27.1",
-        "react": "16.14.0",
-        "react-dom": "16.14.0",
-        "react-query": "3.5.16",
+        "react": "17.0.1",
+        "react-dom": "17.0.1",
+        "react-is": "^17.0.1",
+        "react-query": "3.9.6",
         "react-redux": "7.2.2",
         "react-router": "5.2.0",
         "react-router-dom": "5.2.0",
         "react-router-redux": "4.0.8",
+        "react-transition-group": "4.4.1",
         "redux": "4.0.5",
         "redux-devtools-extension": "2.13.8",
-        "redux-saga": "^1.1.3",
-        "styled-components": "5.2.1",
-        "tsconfig-paths-webpack-plugin": "3.3.0"
+        "redux-saga": "1.1.3",
+        "styled-components": "5.2.1"
     },
     "devDependencies": {
-        "@babel/core": "7.12.10",
-        "@babel/plugin-transform-runtime": "^7.12.10",
-        "@babel/preset-env": "7.12.10",
-        "@babel/preset-react": "7.12.10",
-        "@babel/register": "7.12.10",
-        "@hot-loader/react-dom": "^17.0.1",
+        "@babel/core": "7.12.16",
+        "@babel/plugin-transform-runtime": "7.12.15",
+        "@babel/preset-env": "7.12.16",
+        "@babel/preset-react": "7.12.13",
+        "@babel/register": "7.12.13",
+        "@hot-loader/react-dom": "17.0.1",
+        "@modyqyw/fabric": "1.15.4",
+        "@types/classnames": "2.2.11",
+        "@types/enzyme": "3.10.8",
+        "@types/enzyme-adapter-react-16": "1.0.6",
+        "@types/enzyme-to-json": "1.5.4",
+        "@types/eslint": "7.2.6",
         "@types/history": "4.7.8",
-        "@types/ramda": "0.27.32",
-        "@types/react": "16.14.2",
-        "@types/react-dom": "16.9.10",
+        "@types/jest": "26.0.20",
+        "@types/ramda": "0.27.38",
+        "@types/react": "17.0.2",
+        "@types/react-dom": "17.0.1",
         "@types/react-query": "1.1.2",
-        "@types/react-redux": "7.1.11",
-        "@types/react-router": "5.1.8",
-        "@types/react-router-dom": "4.3.5",
+        "@types/react-redux": "7.1.16",
+        "@types/react-router": "5.1.11",
+        "@types/react-router-dom": "5.1.7",
         "@types/react-router-redux": "5.0.18",
-        "@types/styled-components": "5.1.5",
+        "@types/react-transition-group": "4.4.0",
+        "@types/styled-components": "5.1.7",
+        "@typescript-eslint/parser": "4.15.1",
         "babel-loader": "8.2.2",
-        "braces": "2.3.2",
+        "braces": "3.0.2",
         "circular-dependency-plugin": "5.2.2",
         "classnames": "2.2.6",
-        "compression-webpack-plugin": "3.1.0",
+        "compression-webpack-plugin": "7.1.2",
         "connect-history-api-fallback": "1.6.0",
-        "copy-webpack-plugin": "5.1.2",
-        "core-js": "3.8.1",
-        "cross-env": "5.2.1",
-        "css-loader": "3.6.0",
+        "copy-webpack-plugin": "7.0.0",
+        "core-js": "3.8.3",
+        "cross-env": "7.0.3",
+        "css-loader": "5.0.2",
+        "enzyme": "3.11.0",
+        "enzyme-to-json": "3.6.1",
+        "eslint": "7.20.0",
         "express": "4.17.1",
-        "file-loader": "4.3.0",
-        "fs-extra": "8.1.0",
-        "graceful-fs": "4.2.4",
+        "file-loader": "6.2.0",
+        "fs-extra": "9.1.0",
+        "graceful-fs": "4.2.6",
         "gulp": "4.0.2",
         "gulp-debug": "4.0.0",
-        "html-webpack-plugin": "3.2.0",
-        "html-webpack-template": "~6.2.0",
-        "jest": "24.9.0",
-        "less": "3.12.2",
-        "less-loader": "5.0.0",
+        "html-webpack-plugin": "5.1.0",
+        "jest": "26.6.3",
+        "jest-cli": "26.6.3",
+        "jest-runner-eslint": "0.10.0",
+        "jest-runner-stylelint": "2.3.7",
+        "less": "4.1.1",
+        "less-loader": "8.0.0",
         "lodash": "4.17.20",
-        "mini-css-extract-plugin": "0.12.0",
-        "open": "6.4.0",
+        "mini-css-extract-plugin": "1.3.7",
+        "open": "7.4.1",
+        "prettier": "2.2.1",
         "react-hot-loader": "4.13.0",
         "resolve-url-loader": "3.1.2",
         "run-sequence": "2.2.1",
-        "source-map-loader": "0.2.4",
-        "style-loader": "0.23.1",
-        "svg-url-loader": "3.0.3",
-        "ts-loader": "7.0.5",
-        "typescript": "4.1.3",
-        "webpack": "4.44.2",
-        "webpack-dev-middleware": "3.7.2",
-        "webpack-dev-server": "3.11.0",
+        "source-map-loader": "2.0.1",
+        "style-loader": "2.0.0",
+        "stylelint": "13.10.0",
+        "svg-url-loader": "7.1.1",
+        "ts-jest": "26.5.1",
+        "ts-loader": "8.0.17",
+        "ts-node": "9.1.1",
+        "tsconfig-paths-webpack-plugin": "3.3.0",
+        "typescript": "4.1.5",
+        "webpack": "5.22.0",
+        "webpack-dev-middleware": "4.1.0",
+        "webpack-dev-server": "3.11.2",
         "webpack-env": "0.8.0",
         "webpack-hot-middleware": "2.25.0",
-        "webpack-merge": "4.2.2",
-        "webpack-sources": "1.4.3",
-        "yargs": "13.3.2"
+        "webpack-merge": "5.7.3",
+        "webpack-sources": "2.2.0",
+        "webpack-typings-for-css": "0.6.4",
+        "yargs": "16.2.0"
     }
 }
