@@ -1,9 +1,16 @@
 import React from 'react';
 import { RootApplicationProps } from '@core/components';
 
+export enum ApplicationScope {
+    STANDALONE,
+    MULTIPLE,
+}
 export type ApplicationRegistryEntry<P extends RootApplicationProps> = {
-    component: React.FunctionComponent<P> | React.ComponentClass<P>;
+    Component: React.FunctionComponent<P> | React.ComponentClass<P>;
     props: P;
+    scope: ApplicationScope;
+    name: string;
+    brief?: string;
 };
 
 export type ApplicationRegistry<P extends RootApplicationProps> = { [applicationId: string]: ApplicationRegistryEntry<P> };
