@@ -1,18 +1,20 @@
-import { ApplicationRegistry, ApplicationRegistryEntry, ROOT_APPLICATION_ID } from './API';
+import { ApplicationRegistry, ApplicationRegistryEntry, ApplicationScope, ROOT_APPLICATION_ID } from './API';
 import { RootApplication } from '../components/RootApplication';
 import { StorageType } from '../connectors/API';
 
-const registry: ApplicationRegistry<any> = {
+export const applicationRegistry: ApplicationRegistry<any> = {
     [ROOT_APPLICATION_ID]: {
-        component: RootApplication,
+        Component: RootApplication,
         props: {
             storageType: StorageType.REDUX
-        }
+        },
+        scope: ApplicationScope.STANDALONE,
+        name: "Entry Point Application"
     }
 };
 
-export const getRootApplication = <P>(id: string): ApplicationRegistryEntry<P> => {
-    const application = registry[id];
+export const getRootApplication = <P>(id: string = ROOT_APPLICATION_ID): ApplicationRegistryEntry<P> => {
+    const application = applicationRegistry[id];
 
     if (!application) {
         throw Error(`No root application with ID: ${id}`);
@@ -22,11 +24,11 @@ export const getRootApplication = <P>(id: string): ApplicationRegistryEntry<P> =
 }
 
 export const registerRootApplication = <P>(id: string, entry: ApplicationRegistryEntry<P>): void => {
-    const application = registry[id];
+    const application = applicationRegistry[id];
 
     if (application) {
         throw Error(`Root application with ID "${id}" exists. Please choose another ID`);
     }
 
-    registry[id] = entry;
+    applicationRegistry[id] = entry;
 }
