@@ -1,2 +1,3 @@
 export * from './API';
-export * from './registry';
+export * from './applicationRegistry';
+export * from './manager';
