@@ -3,8 +3,8 @@ import ReactDOM from "react-dom";
 import "core-js/stable";
 import "regenerator-runtime/runtime"
 
-import { App } from './App';
 import { configuration } from "../build/configuration";
+import { App } from './App';
 
 window.onload = () => {
     ReactDOM.render(
